+++ src/main/java/roomescape/controller/ReservationController.java
@@ -12,7 +12,6 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 
 import roomescape.domain.Reservation;
-import roomescape.domain.Time;
 import roomescape.domain.dto.ReservationAddRequest;
 import roomescape.service.ReservationService;
 
@@ -42,27 +41,6 @@ public class ReservationController {
     @DeleteMapping("/reservations/{id}")
     public ResponseEntity<Void> reservationRemove(@PathVariable Long id) {
         reservationService.removeReservation(id);
-
-        return ResponseEntity.noContent().build();
-    }
-
-    @GetMapping("/times")
-    public ResponseEntity<List<Time>> timeList() {
-        List<Time> timeList = reservationService.findTimeList();
-
-        return ResponseEntity.ok().body(timeList);
-    }
-
-    @PostMapping("/times")
-    public ResponseEntity<Time> timeSave(@RequestBody Time time) {
-        Time newTime = reservationService.addTime(time);
-
-        return ResponseEntity.created(URI.create("/times/" + newTime.getId())).body(newTime);
-    }
-
-    @DeleteMapping("/times/{id}")
-    public ResponseEntity<Void> timeRemove(@PathVariable Long id) {
-        reservationService.removeTime(id);
 
         return ResponseEntity.noContent().build();
     }
